improve stest	stability
