+++ src/test/java/stest/tron/wallet/contract/originenergylimit/ContractOriginEnergyLimit001.java
@@ -1,9 +1,11 @@
-package stest.tron.wallet.contract.originEnergyLimit;
+package stest.tron.wallet.contract.originenergylimit;
 
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
+import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Assert;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
@@ -138,4 +140,13 @@ public class ContractOriginEnergyLimit001 {
     Assert.assertEquals(smartContract4.getOriginEnergyLimit(), 99);
   }
 
+  /**
+   * constructor.
+   */
+  @AfterClass
+  public void shutdown() throws InterruptedException {
+    if (channelFull != null) {
+      channelFull.shutdown().awaitTermination(5, TimeUnit.SECONDS);
+    }
+  }
 }
