+++ src/test/java/stest/tron/wallet/contract/trctoken/ContractTrcToken001.java
@@ -1,10 +1,10 @@
-package stest.tron.wallet.contract.trcToken;
-
-import static org.tron.api.GrpcAPI.Return.response_code.CONTRACT_VALIDATE_ERROR;
+package stest.tron.wallet.contract.trctoken;
 
 import com.google.protobuf.ByteString;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
+import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Assert;
@@ -12,20 +12,20 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.tron.api.GrpcAPI;
 import org.tron.api.GrpcAPI.AccountResourceMessage;
 import org.tron.api.WalletGrpc;
 import org.tron.common.crypto.ECKey;
 import org.tron.common.utils.ByteArray;
 import org.tron.common.utils.Utils;
 import org.tron.core.Wallet;
+import org.tron.protos.Protocol.SmartContract;
+import org.tron.protos.Protocol.TransactionInfo;
 import stest.tron.wallet.common.client.Configuration;
 import stest.tron.wallet.common.client.Parameter.CommonConstant;
 import stest.tron.wallet.common.client.utils.PublicMethed;
 
-
 @Slf4j
-public class   ContractTrcToken004 {
+public class ContractTrcToken001 {
 
   private final String testKey002 = Configuration.getByPath("testng.conf")
       .getString("foundationAccount.key1");
@@ -40,9 +40,9 @@ public class   ContractTrcToken004 {
 
   private static final long now = System.currentTimeMillis();
   private static String tokenName = "testAssetIssue_" + Long.toString(now);
-  private static ByteString assetAccountDev = null;
-  private static ByteString assetAccountUser = null;
+  private static ByteString assetAccountId = null;
   private static final long TotalSupply = 1000L;
+  private byte[] transferTokenContractAddress = null;
 
   private String description = Configuration.getByPath("testng.conf")
       .getString("defaultParameter.assetDescription");
@@ -52,10 +52,6 @@ public class   ContractTrcToken004 {
   private ECKey ecKey1 = new ECKey(Utils.getRandom());
   private byte[] dev001Address = ecKey1.getAddress();
   private String dev001Key = ByteArray.toHexString(ecKey1.getPrivKeyBytes());
-
-  private ECKey ecKey2 = new ECKey(Utils.getRandom());
-  private byte[] user001Address = ecKey2.getAddress();
-  private String user001Key = ByteArray.toHexString(ecKey2.getPrivKeyBytes());
 
   @BeforeSuite
   public void beforeSuite() {
@@ -77,44 +73,38 @@ public class   ContractTrcToken004 {
     PublicMethed.printAddress(dev001Key);
   }
 
-
   @Test
   public void deployTransferTokenContract() {
-    Assert.assertTrue(PublicMethed.sendcoin(dev001Address, 1100_000_000L, fromAddress,
+    Assert.assertTrue(PublicMethed.sendcoin(dev001Address, 3100_000_000L, fromAddress,
         testKey002, blockingStubFull));
-    Assert.assertTrue(PublicMethed.sendcoin(user001Address, 1100_000_000L, fromAddress,
-        testKey002, blockingStubFull));
+
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     Assert.assertTrue(PublicMethed.freezeBalanceForReceiver(fromAddress,
-        PublicMethed.getFreezeBalanceCount(dev001Address, dev001Key, 50000L, blockingStubFull),
-        0, 1, ByteString.copyFrom(dev001Address), testKey002, blockingStubFull));
+        PublicMethed.getFreezeBalanceCount(dev001Address, dev001Key, 130000L,
+            blockingStubFull), 0, 1,
+        ByteString.copyFrom(dev001Address), testKey002, blockingStubFull));
+
     Assert.assertTrue(PublicMethed.freezeBalanceForReceiver(fromAddress, 10_000_000L,
-        0, 0, ByteString.copyFrom(dev001Address), testKey002, blockingStubFull));
+        0, 0, ByteString.copyFrom(dev001Address),
+        testKey002, blockingStubFull));
+
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     long start = System.currentTimeMillis() + 2000;
     long end = System.currentTimeMillis() + 1000000000;
-    //user Create a new AssetIssue
-    Assert.assertTrue(PublicMethed.createAssetIssue(user001Address, tokenName, TotalSupply, 1,
-        10000, start, end, 1, description, url, 100000L, 100000L,
-        1L, 1L, user001Key, blockingStubFull));
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    assetAccountUser = PublicMethed.queryAccount(user001Address, blockingStubFull)
-        .getAssetIssuedID();
-    logger.info("The assetAccountUser token name: " + tokenName);
-    logger.info("The assetAccountUser token ID: " + assetAccountUser.toStringUtf8());
 
-    start = System.currentTimeMillis() + 2000;
-    end = System.currentTimeMillis() + 1000000000;
-    //dev Create a new AssetIssue
+    //Create a new AssetIssue success.
     Assert.assertTrue(PublicMethed.createAssetIssue(dev001Address, tokenName, TotalSupply, 1,
         10000, start, end, 1, description, url, 100000L,
         100000L, 1L, 1L, dev001Key, blockingStubFull));
+
     PublicMethed.waitProduceNextBlock(blockingStubFull);
-    assetAccountDev = PublicMethed.queryAccount(dev001Address, blockingStubFull).getAssetIssuedID();
-    logger.info("The assetAccountDev token name: " + tokenName);
-    logger.info("The assetAccountDev token ID: " + assetAccountDev.toStringUtf8());
+
+    assetAccountId = PublicMethed.queryAccount(dev001Address, blockingStubFull).getAssetIssuedID();
+
+    logger.info("The token name: " + tokenName);
+    logger.info("The token ID: " + assetAccountId.toStringUtf8());
 
     //before deploy, check account resource
     AccountResourceMessage accountResource = PublicMethed.getAccountResource(dev001Address,
@@ -123,65 +113,105 @@ public class   ContractTrcToken004 {
     long energyUsage = accountResource.getEnergyUsed();
     long balanceBefore = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
     Long devAssetCountBefore = PublicMethed
-        .getAssetIssueValue(dev001Address, assetAccountDev, blockingStubFull);
-    Long userAssetCountBefore = PublicMethed
-        .getAssetIssueValue(dev001Address, assetAccountUser, blockingStubFull);
+        .getAssetIssueValue(dev001Address, assetAccountId, blockingStubFull);
 
     logger.info("before energyLimit is " + Long.toString(energyLimit));
     logger.info("before energyUsage is " + Long.toString(energyUsage));
     logger.info("before balanceBefore is " + Long.toString(balanceBefore));
-    logger.info("before dev has AssetId: " + assetAccountDev.toStringUtf8()
+    logger.info("before AssetId: " + assetAccountId.toStringUtf8()
         + ", devAssetCountBefore: " + devAssetCountBefore);
-    logger.info("before dev has AssetId: " + assetAccountUser.toStringUtf8()
-        + ", userAssetCountBefore: " + userAssetCountBefore);
 
     String contractName = "transferTokenContract";
-    String code = "608060405260e2806100126000396000f300608060405260043610603e5763ffffffff7c01000000"
-        + "000000000000000000000000000000000000000000000000006000350416633be9ece781146043575b600080"
-        + "fd5b606873ffffffffffffffffffffffffffffffffffffffff60043516602435604435606a565b005b604051"
-        + "73ffffffffffffffffffffffffffffffffffffffff84169082156108fc029083908590600081818185878a8a"
-        + "d094505050505015801560b0573d6000803e3d6000fd5b505050505600a165627a7a723058200ba246bdb58b"
-        + "e0f221ad07e1b19de843ab541150b329ddd01558c2f1cefe1e270029";
-    String abi = "[{\"constant\":false,\"inputs\":[{\"name\":\"toAddress\",\"type\":\"address\"},"
-        + "{\"name\":\"id\",\"type\":\"trcToken\"},{\"name\":\"amount\",\"type\":\"uint256\"}],"
-        + "\"name\":\"TransferTokenTo\",\"outputs\":[],\"payable\":true,\"stateMutability\":"
-        + "\"payable\",\"type\":\"function\"},{\"inputs\":[],\"payable\":true,\"stateMutability\":"
-        + "\"payable\",\"type\":\"constructor\"}]";
+    String code = Configuration.getByPath("testng.conf")
+        .getString("code.code_ContractTrcToken001_transferTokenContract");
+    String abi = Configuration.getByPath("testng.conf")
+        .getString("abi.abi_ContractTrcToken001_transferTokenContract");
 
-    // the tokenId is not exist
-    String fakeTokenId = assetAccountDev.toStringUtf8();
-    Long fakeTokenValue = devAssetCountBefore + 100;
+    String tokenId = assetAccountId.toStringUtf8();
+    long tokenValue = 100;
+    long callValue = 5;
 
-    GrpcAPI.Return response = PublicMethed
-        .deployContractAndGetResponse(contractName, abi, code, "",
-            maxFeeLimit, 0L, 0, 10000,
-            fakeTokenId, fakeTokenValue, null, dev001Key,
-            dev001Address, blockingStubFull);
+    final String transferTokenTxid = PublicMethed
+        .deployContractAndGetTransactionInfoById(contractName, abi, code, "", maxFeeLimit,
+            callValue, 0, 10000, tokenId, tokenValue,
+            null, dev001Key, dev001Address, blockingStubFull);
 
-    Assert.assertFalse(response.getResult());
-    Assert.assertEquals(CONTRACT_VALIDATE_ERROR, response.getCode());
-    Assert.assertEquals("contract validate error : assetBalance is not sufficient.",
-        response.getMessage().toStringUtf8());
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     accountResource = PublicMethed.getAccountResource(dev001Address, blockingStubFull);
     energyLimit = accountResource.getEnergyLimit();
     energyUsage = accountResource.getEnergyUsed();
     long balanceAfter = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
     Long devAssetCountAfter = PublicMethed
-        .getAssetIssueValue(dev001Address, assetAccountDev, blockingStubFull);
-    Long userAssetCountAfter = PublicMethed
-        .getAssetIssueValue(dev001Address, assetAccountUser, blockingStubFull);
+        .getAssetIssueValue(dev001Address, assetAccountId, blockingStubFull);
 
     logger.info("after energyLimit is " + Long.toString(energyLimit));
     logger.info("after energyUsage is " + Long.toString(energyUsage));
     logger.info("after balanceAfter is " + Long.toString(balanceAfter));
-    logger.info("after dev has AssetId: " + assetAccountDev.toStringUtf8()
+    logger.info("after AssetId: " + assetAccountId.toStringUtf8()
         + ", devAssetCountAfter: " + devAssetCountAfter);
-    logger.info("after user has AssetId: " + assetAccountDev.toStringUtf8()
-        + ", userAssetCountAfter: " + userAssetCountAfter);
 
-    Assert.assertEquals(devAssetCountBefore, devAssetCountAfter);
-    Assert.assertEquals(userAssetCountBefore, userAssetCountAfter);
+    Optional<TransactionInfo> infoById = PublicMethed
+        .getTransactionInfoById(transferTokenTxid, blockingStubFull);
+
+    if (transferTokenTxid == null || infoById.get().getResultValue() != 0) {
+      Assert.fail("deploy transaction failed with message: " + infoById.get().getResMessage());
+    }
+
+    transferTokenContractAddress = infoById.get().getContractAddress().toByteArray();
+    SmartContract smartContract = PublicMethed.getContract(transferTokenContractAddress,
+        blockingStubFull);
+    Assert.assertNotNull(smartContract.getAbi());
+
+    Assert.assertTrue(PublicMethed.transferAsset(transferTokenContractAddress,
+        assetAccountId.toByteArray(), 100L, dev001Address, dev001Key, blockingStubFull));
+
+    Long contractAssetCount = PublicMethed.getAssetIssueValue(transferTokenContractAddress,
+        assetAccountId, blockingStubFull);
+    logger.info("Contract has AssetId: " + assetAccountId.toStringUtf8() + ", Count: "
+        + contractAssetCount);
+
+    Assert.assertEquals(Long.valueOf(tokenValue),
+        Long.valueOf(devAssetCountBefore - devAssetCountAfter));
+    Assert.assertEquals(Long.valueOf(100L + tokenValue), contractAssetCount);
+
+    // get and verify the msg.value and msg.id
+    Long transferAssetBefore = PublicMethed
+        .getAssetIssueValue(transferTokenContractAddress, assetAccountId,
+            blockingStubFull);
+    logger.info("before trigger, transferTokenContractAddress has AssetId "
+        + assetAccountId.toStringUtf8() + ", Count is " + transferAssetBefore);
+
+    final String triggerTxid = PublicMethed.triggerContract(transferTokenContractAddress,
+        "getResultInCon()", "#", false, 0,
+        1000000000L, "0", 0, dev001Address, dev001Key,
+        blockingStubFull);
+
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+
+    infoById = PublicMethed
+        .getTransactionInfoById(triggerTxid, blockingStubFull);
+    if (infoById.get().getResultValue() != 0) {
+      Assert.fail("transaction failed with message: " + infoById.get().getResMessage());
+    }
+
+    logger.info("The msg value: " + PublicMethed
+        .getStrings(infoById.get().getContractResult(0).toByteArray()));
+
+    List<String> retList = PublicMethed
+        .getStrings(infoById.get().getContractResult(0).toByteArray());
+
+    Long msgId = ByteArray.toLong(ByteArray.fromHexString(retList.get(0)));
+    Long msgTokenValue = ByteArray.toLong(ByteArray.fromHexString(retList.get(1)));
+    Long msgCallValue = ByteArray.toLong(ByteArray.fromHexString(retList.get(2)));
+
+    logger.info("msgId: " + msgId);
+    logger.info("msgTokenValue: " + msgTokenValue);
+    logger.info("msgCallValue: " + msgCallValue);
+
+    Assert.assertEquals(msgId.toString(), tokenId);
+    Assert.assertEquals(Long.valueOf(msgTokenValue), Long.valueOf(tokenValue));
+    Assert.assertEquals(Long.valueOf(msgCallValue), Long.valueOf(callValue));
 
     PublicMethed.unFreezeBalance(fromAddress, testKey002, 1,
         dev001Address, blockingStubFull);
