+++ src/test/java/stest/tron/wallet/contract/trctoken/ContractTrcToken002.java
@@ -1,10 +1,9 @@
-package stest.tron.wallet.contract.trcToken;
-
-import static org.tron.api.GrpcAPI.Return.response_code.CONTRACT_VALIDATE_ERROR;
+package stest.tron.wallet.contract.trctoken;
 
 import com.google.protobuf.ByteString;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
+import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
@@ -13,7 +12,6 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.tron.api.GrpcAPI;
 import org.tron.api.GrpcAPI.AccountResourceMessage;
 import org.tron.api.WalletGrpc;
 import org.tron.common.crypto.ECKey;
@@ -27,7 +25,7 @@ import stest.tron.wallet.common.client.Parameter.CommonConstant;
 import stest.tron.wallet.common.client.utils.PublicMethed;
 
 @Slf4j
-public class ContractTrcToken006 {
+public class ContractTrcToken002 {
 
   private final String testKey002 = Configuration.getByPath("testng.conf")
       .getString("foundationAccount.key1");
@@ -80,12 +78,12 @@ public class ContractTrcToken006 {
     PublicMethed.printAddress(user001Key);
   }
 
-
-  @Test
+  @Test(enabled = true)
   public void deployTransferTokenContract() {
-    Assert.assertTrue(PublicMethed.sendcoin(dev001Address, 1100_000_000L, fromAddress,
+
+    Assert.assertTrue(PublicMethed.sendcoin(dev001Address, 3100_000_000L, fromAddress,
         testKey002, blockingStubFull));
-    Assert.assertTrue(PublicMethed.sendcoin(user001Address, 1_000_000L, fromAddress,
+    Assert.assertTrue(PublicMethed.sendcoin(user001Address, 300_000_000L, fromAddress,
         testKey002, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
@@ -94,17 +92,18 @@ public class ContractTrcToken006 {
             blockingStubFull), 0, 1,
         ByteString.copyFrom(dev001Address), testKey002, blockingStubFull));
     Assert.assertTrue(PublicMethed.freezeBalanceForReceiver(fromAddress, 10_000_000L,
-        0, 0, ByteString.copyFrom(dev001Address), testKey002, blockingStubFull));
+        0, 0, ByteString.copyFrom(dev001Address),
+        testKey002, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     long start = System.currentTimeMillis() + 2000;
     long end = System.currentTimeMillis() + 1000000000;
     //Create a new AssetIssue success.
     Assert.assertTrue(PublicMethed.createAssetIssue(dev001Address, tokenName, TotalSupply, 1,
-        10000, start, end, 1, description, url, 100000L, 100000L,
-        1L, 1L, dev001Key, blockingStubFull));
-    assetAccountId = PublicMethed
-        .queryAccount(dev001Address, blockingStubFull).getAssetIssuedID();
+        10000, start, end, 1, description, url, 100000L,
+        100000L, 1L, 1L, dev001Key, blockingStubFull));
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    assetAccountId = PublicMethed.queryAccount(dev001Address, blockingStubFull).getAssetIssuedID();
     logger.info("The token name: " + tokenName);
     logger.info("The token ID: " + assetAccountId.toStringUtf8());
 
@@ -114,8 +113,8 @@ public class ContractTrcToken006 {
     long energyLimit = accountResource.getEnergyLimit();
     long energyUsage = accountResource.getEnergyUsed();
     long balanceBefore = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
-    Long devAssetCountBefore = PublicMethed
-        .getAssetIssueValue(dev001Address, assetAccountId, blockingStubFull);
+    Long devAssetCountBefore = PublicMethed.getAssetIssueValue(
+        dev001Address, assetAccountId, blockingStubFull);
 
     logger.info("before energyLimit is " + Long.toString(energyLimit));
     logger.info("before energyUsage is " + Long.toString(energyUsage));
@@ -125,9 +124,10 @@ public class ContractTrcToken006 {
 
     String contractName = "transferTokenContract";
     String code = Configuration.getByPath("testng.conf")
-        .getString("code.code_ContractTrcToken006_transferTokenContract");
+        .getString("code.code_ContractTrcToken002_transferTokenContract");
     String abi = Configuration.getByPath("testng.conf")
-        .getString("abi.abi_ContractTrcToken006_transferTokenContract");
+        .getString("abi.abi_ContractTrcToken002_transferTokenContract");
+
     String tokenId = assetAccountId.toStringUtf8();
     long tokenValue = 100;
     long callValue = 0;
@@ -155,8 +155,8 @@ public class ContractTrcToken006 {
     energyLimit = accountResource.getEnergyLimit();
     energyUsage = accountResource.getEnergyUsed();
     long balanceAfter = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
-    Long devAssetCountAfter = PublicMethed
-        .getAssetIssueValue(dev001Address, assetAccountId, blockingStubFull);
+    Long devAssetCountAfter = PublicMethed.getAssetIssueValue(
+        dev001Address, assetAccountId, blockingStubFull);
 
     logger.info("after energyLimit is " + Long.toString(energyLimit));
     logger.info("after energyUsage is " + Long.toString(energyUsage));
@@ -170,8 +170,8 @@ public class ContractTrcToken006 {
 
     Long contractAssetCount = PublicMethed.getAssetIssueValue(transferTokenContractAddress,
         assetAccountId, blockingStubFull);
-    logger.info("Contract has AssetId: " + assetAccountId.toStringUtf8() + ", Count: "
-        + contractAssetCount);
+    logger.info("Contract has AssetId: " + assetAccountId.toStringUtf8()
+        + ", Count: " + contractAssetCount);
 
     Assert.assertEquals(Long.valueOf(100), Long.valueOf(devAssetCountBefore - devAssetCountAfter));
     Assert.assertEquals(Long.valueOf(200), contractAssetCount);
@@ -180,6 +180,9 @@ public class ContractTrcToken006 {
         PublicMethed.getFreezeBalanceCount(user001Address, user001Key, 50000L,
             blockingStubFull), 0, 1,
         ByteString.copyFrom(user001Address), testKey002, blockingStubFull));
+
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+
     Assert.assertTrue(PublicMethed.transferAsset(user001Address,
         assetAccountId.toByteArray(), 10L, dev001Address, dev001Key, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
@@ -205,31 +208,24 @@ public class ContractTrcToken006 {
     logger.info("before trigger, userBalanceBefore is " + Long.toString(userBalanceBefore));
 
     Long transferAssetBefore = PublicMethed
-        .getAssetIssueValue(transferTokenContractAddress, assetAccountId,
-        blockingStubFull);
+        .getAssetIssueValue(transferTokenContractAddress, assetAccountId, blockingStubFull);
     logger.info("before trigger, transferTokenContractAddress has AssetId "
         + assetAccountId.toStringUtf8() + ", Count is " + transferAssetBefore);
 
-    Long userAssetId = PublicMethed
-        .getAssetIssueValue(user001Address, ByteString.copyFromUtf8(tokenId),
-        blockingStubFull);
-    logger.info("before userAssetId has AssetId "
-        + tokenId + ", Count is " + userAssetId);
+    PublicMethed
+        .sendcoin(transferTokenContractAddress, 5000000, fromAddress, testKey002, blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     tokenId = assetAccountId.toStringUtf8();
-    tokenValue = 100;
-    callValue = 5;
+    tokenValue = 10;
+    callValue = 0;
 
-    GrpcAPI.Return response = PublicMethed
-        .triggerContractAndGetResponse(transferTokenContractAddress,
-            "msgTokenValueAndTokenIdTest()", "#", false, callValue,
-            1000000000L, tokenId, tokenValue, user001Address, user001Key,
-            blockingStubFull);
+    final String triggerTxid = PublicMethed.triggerContract(transferTokenContractAddress,
+        "msgTokenValueAndTokenIdTest()", "#", false, callValue,
+        1000000000L, tokenId, tokenValue, user001Address, user001Key,
+        blockingStubFull);
 
-    Assert.assertFalse(response.getResult());
-    Assert.assertEquals(CONTRACT_VALIDATE_ERROR, response.getCode());
-    Assert.assertEquals("contract validate error : assetBalance is not sufficient.",
-        response.getMessage().toStringUtf8());
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     accountResource = PublicMethed.getAccountResource(dev001Address, blockingStubFull);
     long devEnergyLimitAfter = accountResource.getEnergyLimit();
@@ -243,12 +239,36 @@ public class ContractTrcToken006 {
     accountResource = PublicMethed.getAccountResource(user001Address, blockingStubFull);
     long userEnergyLimitAfter = accountResource.getEnergyLimit();
     long userEnergyUsageAfter = accountResource.getEnergyUsed();
-    long userBalanceAfter = PublicMethed.queryAccount(user001Address, blockingStubFull)
-        .getBalance();
+    long userBalanceAfter = PublicMethed.queryAccount(user001Address,
+        blockingStubFull).getBalance();
 
     logger.info("after trigger, userEnergyLimitAfter is " + Long.toString(userEnergyLimitAfter));
     logger.info("after trigger, userEnergyUsageAfter is " + Long.toString(userEnergyUsageAfter));
     logger.info("after trigger, userBalanceAfter is " + Long.toString(userBalanceAfter));
+
+    infoById = PublicMethed
+        .getTransactionInfoById(triggerTxid, blockingStubFull);
+    if (triggerTxid == null || infoById.get().getResultValue() != 0) {
+      Assert.fail("transaction failed with message: " + infoById.get().getResMessage());
+    }
+
+    logger.info("the value: "
+        + PublicMethed.getStrings(infoById.get().getContractResult(0).toByteArray()));
+
+    List<String> retList = PublicMethed
+        .getStrings(infoById.get().getContractResult(0).toByteArray());
+
+    Long msgId = ByteArray.toLong(ByteArray.fromHexString(retList.get(0)));
+    Long msgTokenValue = ByteArray.toLong(ByteArray.fromHexString(retList.get(1)));
+    Long msgCallValue = ByteArray.toLong(ByteArray.fromHexString(retList.get(2)));
+
+    logger.info("msgId: " + msgId);
+    logger.info("msgTokenValue: " + msgTokenValue);
+    logger.info("msgCallValue: " + msgCallValue);
+
+    Assert.assertEquals(msgId.toString(), tokenId);
+    Assert.assertEquals(Long.valueOf(msgTokenValue), Long.valueOf(tokenValue));
+    Assert.assertEquals(Long.valueOf(msgCallValue), Long.valueOf(callValue));
 
     PublicMethed.unFreezeBalance(fromAddress, testKey002, 1,
         dev001Address, blockingStubFull);
