+++ src/test/java/stest/tron/wallet/contract/trctoken/ContractTrcToken012.java
@@ -1,9 +1,8 @@
-package stest.tron.wallet.contract.trcToken;
+package stest.tron.wallet.contract.trctoken;
 
 import com.google.protobuf.ByteString;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
@@ -22,10 +21,12 @@ import org.tron.protos.Protocol.SmartContract;
 import org.tron.protos.Protocol.TransactionInfo;
 import stest.tron.wallet.common.client.Configuration;
 import stest.tron.wallet.common.client.Parameter.CommonConstant;
+import stest.tron.wallet.common.client.utils.Base58;
 import stest.tron.wallet.common.client.utils.PublicMethed;
 
 @Slf4j
-public class ContractTrcToken002 {
+public class ContractTrcToken012 {
+
   private final String testKey002 = Configuration.getByPath("testng.conf")
       .getString("foundationAccount.key1");
   private final byte[] fromAddress = PublicMethed.getFinalAddress(testKey002);
@@ -72,37 +73,34 @@ public class ContractTrcToken002 {
         .usePlaintext(true)
         .build();
     blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    PublicMethed.printAddress(dev001Key);
-    PublicMethed.printAddress(user001Key);
   }
 
   @Test(enabled = true)
-  public void deployTransferTokenContract() {
-
-    Assert.assertTrue(PublicMethed.sendcoin(dev001Address, 3100_000_000L, fromAddress,
+  public void test01DeployTransferTokenContract() {
+    Assert.assertTrue(PublicMethed.sendcoin(dev001Address, 15048_000_000L, fromAddress,
         testKey002, blockingStubFull));
-    Assert.assertTrue(PublicMethed.sendcoin(user001Address, 300_000_000L, fromAddress,
+    Assert.assertTrue(PublicMethed.sendcoin(user001Address, 14048_000_000L, fromAddress,
         testKey002, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     Assert.assertTrue(PublicMethed.freezeBalanceForReceiver(fromAddress,
-        PublicMethed.getFreezeBalanceCount(dev001Address, dev001Key, 70000L,
+        PublicMethed.getFreezeBalanceCount(dev001Address, dev001Key, 170000L,
             blockingStubFull), 0, 1,
         ByteString.copyFrom(dev001Address), testKey002, blockingStubFull));
+
     Assert.assertTrue(PublicMethed.freezeBalanceForReceiver(fromAddress, 10_000_000L,
-        0, 0, ByteString.copyFrom(dev001Address),
-        testKey002, blockingStubFull));
+        0, 0, ByteString.copyFrom(dev001Address), testKey002, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     long start = System.currentTimeMillis() + 2000;
     long end = System.currentTimeMillis() + 1000000000;
     //Create a new AssetIssue success.
     Assert.assertTrue(PublicMethed.createAssetIssue(dev001Address, tokenName, TotalSupply, 1,
-        10000, start, end, 1, description, url, 100000L,
-        100000L, 1L, 1L, dev001Key, blockingStubFull));
+        10000, start, end, 1, description, url, 100000L, 100000L,
+        1L, 1L, dev001Key, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
-    assetAccountId = PublicMethed.queryAccount(dev001Address, blockingStubFull).getAssetIssuedID();
+    assetAccountId = PublicMethed
+        .queryAccount(dev001Address, blockingStubFull).getAssetIssuedID();
     logger.info("The token name: " + tokenName);
     logger.info("The token ID: " + assetAccountId.toStringUtf8());
 
@@ -112,8 +110,8 @@ public class ContractTrcToken002 {
     long energyLimit = accountResource.getEnergyLimit();
     long energyUsage = accountResource.getEnergyUsed();
     long balanceBefore = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
-    Long devAssetCountBefore = PublicMethed.getAssetIssueValue(
-        dev001Address, assetAccountId, blockingStubFull);
+    Long devAssetCountBefore = PublicMethed
+        .getAssetIssueValue(dev001Address, assetAccountId, blockingStubFull);
 
     logger.info("before energyLimit is " + Long.toString(energyLimit));
     logger.info("before energyUsage is " + Long.toString(energyUsage));
@@ -123,20 +121,29 @@ public class ContractTrcToken002 {
 
     String contractName = "transferTokenContract";
     String code = Configuration.getByPath("testng.conf")
-        .getString("code.code_ContractTrcToken002_transferTokenContract");
+        .getString("code.code_ContractTrcToken012_transferTokenContract");
     String abi = Configuration.getByPath("testng.conf")
-        .getString("abi.abi_ContractTrcToken002_transferTokenContract");
+        .getString("abi.abi_ContractTrcToken012_transferTokenContract");
 
-    String tokenId = assetAccountId.toStringUtf8();
-    long tokenValue = 100;
-    long callValue = 0;
-
-    String transferTokenTxid = PublicMethed
+    final String transferTokenTxid = PublicMethed
         .deployContractAndGetTransactionInfoById(contractName, abi, code, "",
-            maxFeeLimit, callValue, 0, 10000,
-            tokenId, tokenValue, null, dev001Key,
+            maxFeeLimit, 0L, 0, 10000,
+            assetAccountId.toStringUtf8(), 100, null, dev001Key,
             dev001Address, blockingStubFull);
     PublicMethed.waitProduceNextBlock(blockingStubFull);
+
+    accountResource = PublicMethed.getAccountResource(dev001Address, blockingStubFull);
+    energyLimit = accountResource.getEnergyLimit();
+    energyUsage = accountResource.getEnergyUsed();
+    long balanceAfter = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
+    Long devAssetCountAfter = PublicMethed
+        .getAssetIssueValue(dev001Address, assetAccountId, blockingStubFull);
+
+    logger.info("after energyLimit is " + Long.toString(energyLimit));
+    logger.info("after energyUsage is " + Long.toString(energyUsage));
+    logger.info("after balanceAfter is " + Long.toString(balanceAfter));
+    logger.info("after AssetId: " + assetAccountId.toStringUtf8()
+        + ", devAssetCountAfter: " + devAssetCountAfter);
 
     Optional<TransactionInfo> infoById = PublicMethed
         .getTransactionInfoById(transferTokenTxid, blockingStubFull);
@@ -150,43 +157,31 @@ public class ContractTrcToken002 {
         blockingStubFull);
     Assert.assertNotNull(smartContract.getAbi());
 
-    accountResource = PublicMethed.getAccountResource(dev001Address, blockingStubFull);
-    energyLimit = accountResource.getEnergyLimit();
-    energyUsage = accountResource.getEnergyUsed();
-    long balanceAfter = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
-    Long devAssetCountAfter = PublicMethed.getAssetIssueValue(
-        dev001Address, assetAccountId, blockingStubFull);
-
-    logger.info("after energyLimit is " + Long.toString(energyLimit));
-    logger.info("after energyUsage is " + Long.toString(energyUsage));
-    logger.info("after balanceAfter is " + Long.toString(balanceAfter));
-    logger.info("after AssetId: " + assetAccountId.toStringUtf8()
-        + ", devAssetCountAfter: " + devAssetCountAfter);
-
     Assert.assertTrue(PublicMethed.transferAsset(transferTokenContractAddress,
         assetAccountId.toByteArray(), 100L, dev001Address, dev001Key, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     Long contractAssetCount = PublicMethed.getAssetIssueValue(transferTokenContractAddress,
         assetAccountId, blockingStubFull);
-    logger.info("Contract has AssetId: " + assetAccountId.toStringUtf8()
-        + ", Count: " + contractAssetCount);
+    logger.info("Contract has AssetId: " + assetAccountId.toStringUtf8() + ", Count: "
+        + contractAssetCount);
 
     Assert.assertEquals(Long.valueOf(100), Long.valueOf(devAssetCountBefore - devAssetCountAfter));
     Assert.assertEquals(Long.valueOf(200), contractAssetCount);
+  }
 
+
+  @Test(enabled = true)
+  public void test02TriggerContract() {
     Assert.assertTrue(PublicMethed.freezeBalanceForReceiver(fromAddress,
         PublicMethed.getFreezeBalanceCount(user001Address, user001Key, 50000L,
             blockingStubFull), 0, 1,
         ByteString.copyFrom(user001Address), testKey002, blockingStubFull));
-
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-
     Assert.assertTrue(PublicMethed.transferAsset(user001Address,
         assetAccountId.toByteArray(), 10L, dev001Address, dev001Key, blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
-    accountResource = PublicMethed.getAccountResource(dev001Address,
+    AccountResourceMessage accountResource = PublicMethed.getAccountResource(dev001Address,
         blockingStubFull);
     long devEnergyLimitBefore = accountResource.getEnergyLimit();
     long devEnergyUsageBefore = accountResource.getEnergyUsed();
@@ -207,23 +202,27 @@ public class ContractTrcToken002 {
     logger.info("before trigger, userBalanceBefore is " + Long.toString(userBalanceBefore));
 
     Long transferAssetBefore = PublicMethed
-        .getAssetIssueValue(transferTokenContractAddress, assetAccountId, blockingStubFull);
+        .getAssetIssueValue(transferTokenContractAddress, assetAccountId,
+            blockingStubFull);
     logger.info("before trigger, transferTokenContractAddress has AssetId "
         + assetAccountId.toStringUtf8() + ", Count is " + transferAssetBefore);
 
-    PublicMethed
-        .sendcoin(transferTokenContractAddress, 5000000, fromAddress, testKey002, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    Long receiveAssetBefore = PublicMethed.getAssetIssueValue(dev001Address, assetAccountId,
+        blockingStubFull);
+    logger.info("before trigger, receiveTokenContractAddress has AssetId "
+        + assetAccountId.toStringUtf8() + ", Count is " + receiveAssetBefore);
 
-    tokenId = assetAccountId.toStringUtf8();
-    tokenValue = 10;
-    callValue = 0;
+    String tokenId = assetAccountId.toStringUtf8();
+    Long tokenValue = Long.valueOf(1);
+    Long callValue = Long.valueOf(0);
+
+    String param = "\"" + Base58.encode58Check(dev001Address)
+        + "\",\"" + tokenValue + "\"," + tokenId;
 
     final String triggerTxid = PublicMethed.triggerContract(transferTokenContractAddress,
-        "msgTokenValueAndTokenIdTest()", "#", false, callValue,
-        1000000000L, tokenId, tokenValue, user001Address, user001Key,
+        "transferTokenTest(address,uint256,trcToken)", param, false, callValue,
+        1000000000L, assetAccountId.toStringUtf8(), 2, user001Address, user001Key,
         blockingStubFull);
-
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
     accountResource = PublicMethed.getAccountResource(dev001Address, blockingStubFull);
@@ -238,38 +237,123 @@ public class ContractTrcToken002 {
     accountResource = PublicMethed.getAccountResource(user001Address, blockingStubFull);
     long userEnergyLimitAfter = accountResource.getEnergyLimit();
     long userEnergyUsageAfter = accountResource.getEnergyUsed();
-    long userBalanceAfter = PublicMethed.queryAccount(user001Address,
-        blockingStubFull).getBalance();
+    long userBalanceAfter = PublicMethed.queryAccount(user001Address, blockingStubFull)
+        .getBalance();
 
     logger.info("after trigger, userEnergyLimitAfter is " + Long.toString(userEnergyLimitAfter));
     logger.info("after trigger, userEnergyUsageAfter is " + Long.toString(userEnergyUsageAfter));
     logger.info("after trigger, userBalanceAfter is " + Long.toString(userBalanceAfter));
 
-
-    infoById = PublicMethed
+    Optional<TransactionInfo> infoById = PublicMethed
         .getTransactionInfoById(triggerTxid, blockingStubFull);
     if (triggerTxid == null || infoById.get().getResultValue() != 0) {
       Assert.fail("transaction failed with message: " + infoById.get().getResMessage());
     }
 
-    logger.info("the value: "
-        + PublicMethed.getStrings(infoById.get().getContractResult(0).toByteArray()));
+    SmartContract smartContract = PublicMethed.getContract(infoById.get().getContractAddress()
+        .toByteArray(), blockingStubFull);
 
-    List<String> retList = PublicMethed
-        .getStrings(infoById.get().getContractResult(0).toByteArray());
+    Long transferAssetAfter = PublicMethed.getAssetIssueValue(transferTokenContractAddress,
+        assetAccountId, blockingStubFull);
+    logger.info("after trigger, transferTokenContractAddress has AssetId "
+        + assetAccountId.toStringUtf8() + ", transferAssetAfter is " + transferAssetAfter);
 
-    Long msgId = ByteArray.toLong(ByteArray.fromHexString(retList.get(0)));
-    Long msgTokenValue = ByteArray.toLong(ByteArray.fromHexString(retList.get(1)));
-    Long msgCallValue = ByteArray.toLong(ByteArray.fromHexString(retList.get(2)));
+    Long receiveAssetAfter = PublicMethed.getAssetIssueValue(dev001Address,
+        assetAccountId, blockingStubFull);
+    logger.info("after trigger, resultContractAddress has AssetId "
+        + assetAccountId.toStringUtf8() + ", receiveAssetAfter is " + receiveAssetAfter);
 
-    logger.info("msgId: " + msgId);
-    logger.info("msgTokenValue: " + msgTokenValue);
-    logger.info("msgCallValue: " + msgCallValue);
+    long consumeUserPercent = smartContract.getConsumeUserResourcePercent();
+    logger.info("ConsumeURPercent: " + consumeUserPercent);
 
-    Assert.assertEquals(msgId.toString(), tokenId);
-    Assert.assertEquals(Long.valueOf(msgTokenValue), Long.valueOf(tokenValue));
-    Assert.assertEquals(Long.valueOf(msgCallValue), Long.valueOf(callValue));
+    Assert.assertEquals(receiveAssetAfter - receiveAssetBefore,
+        transferAssetBefore + 2L - transferAssetAfter);
+  }
 
+  @Test(enabled = true)
+  public void test03TriggerTokenBalanceContract() {
+    Assert.assertTrue(PublicMethed.freezeBalanceGetEnergy(user001Address, 1000_000_000L,
+        0, 1, user001Key, blockingStubFull));
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+
+    AccountResourceMessage accountResource = PublicMethed.getAccountResource(dev001Address,
+        blockingStubFull);
+    long devEnergyLimitBefore = accountResource.getEnergyLimit();
+    long devEnergyUsageBefore = accountResource.getEnergyUsed();
+    long devBalanceBefore = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
+
+    logger.info("before trigger, dev energy limit is " + Long.toString(devEnergyLimitBefore));
+    logger.info("before trigger, dev energy usage is " + Long.toString(devEnergyUsageBefore));
+    logger.info("before trigger, dev balance is " + Long.toString(devBalanceBefore));
+
+    accountResource = PublicMethed.getAccountResource(user001Address, blockingStubFull);
+    long userEnergyLimitBefore = accountResource.getEnergyLimit();
+    long userEnergyUsageBefore = accountResource.getEnergyUsed();
+    long userBalanceBefore = PublicMethed.queryAccount(user001Address,
+        blockingStubFull).getBalance();
+
+    logger.info("before trigger, user energy limit is " + Long.toString(userEnergyLimitBefore));
+    logger.info("before trigger, user energy usage is " + Long.toString(userEnergyUsageBefore));
+    logger.info("before trigger, user balance is " + Long.toString(userBalanceBefore));
+
+    String param = "\"" + Base58.encode58Check(dev001Address) + "\",\""
+        + assetAccountId.toStringUtf8() + "\"";
+
+    final String triggerTxid = PublicMethed.triggerContract(transferTokenContractAddress,
+        "getTokenBalnce(address,trcToken)",
+        param, false, 0, 1000000000L, user001Address,
+        user001Key, blockingStubFull);
+
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+
+    accountResource = PublicMethed.getAccountResource(dev001Address, blockingStubFull);
+    long devEnergyLimitAfter = accountResource.getEnergyLimit();
+    long devEnergyUsageAfter = accountResource.getEnergyUsed();
+    long devBalanceAfter = PublicMethed.queryAccount(dev001Key, blockingStubFull).getBalance();
+
+    logger.info("after trigger, devEnergyLimitAfter is " + Long.toString(devEnergyLimitAfter));
+    logger.info("after trigger, devEnergyUsageAfter is " + Long.toString(devEnergyUsageAfter));
+    logger.info("after trigger, devBalanceAfter is " + Long.toString(devBalanceAfter));
+
+    accountResource = PublicMethed.getAccountResource(user001Address, blockingStubFull);
+    long userEnergyLimitAfter = accountResource.getEnergyLimit();
+    long userEnergyUsageAfter = accountResource.getEnergyUsed();
+    long userBalanceAfter = PublicMethed.queryAccount(user001Address, blockingStubFull)
+        .getBalance();
+
+    logger.info("after trigger, userEnergyLimitAfter is " + Long.toString(userEnergyLimitAfter));
+    logger.info("after trigger, userEnergyUsageAfter is " + Long.toString(userEnergyUsageAfter));
+    logger.info("after trigger, userBalanceAfter is " + Long.toString(userBalanceAfter));
+
+    Optional<TransactionInfo> infoById = PublicMethed.getTransactionInfoById(triggerTxid,
+        blockingStubFull);
+    if (triggerTxid == null || infoById.get().getResultValue() != 0) {
+      Assert.fail("transaction failed with message: " + infoById.get().getResMessage());
+    }
+
+    SmartContract smartContract = PublicMethed.getContract(infoById.get().getContractAddress()
+        .toByteArray(), blockingStubFull);
+
+    long consumeUserPercent = smartContract.getConsumeUserResourcePercent();
+    logger.info("ConsumeURPercent: " + consumeUserPercent);
+
+    infoById = PublicMethed.getTransactionInfoById(triggerTxid, blockingStubFull);
+
+    if (infoById.get().getResultValue() != 0) {
+      Assert.fail("transaction failed with message: " + infoById.get().getResMessage());
+    }
+    logger.info("the receivercontract token: " + ByteArray
+        .toLong(infoById.get().getContractResult(0).toByteArray()));
+    Long assetIssueCount = PublicMethed
+        .getAssetIssueValue(dev001Address, assetAccountId,
+            blockingStubFull);
+
+    logger.info("the receivercontract token(getaccount): " + assetIssueCount);
+    Assert.assertTrue(assetIssueCount == ByteArray
+        .toLong(ByteArray.fromHexString(
+            ByteArray.toHexString(infoById.get().getContractResult(0).toByteArray()))));
+
+    // unfreeze resource
     PublicMethed.unFreezeBalance(fromAddress, testKey002, 1,
         dev001Address, blockingStubFull);
     PublicMethed.unFreezeBalance(fromAddress, testKey002, 0,
